requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-text-splitters
langchain-core
chromadb
sentence-transformers

